scripts folder & seeds
