+++ package.json
@@ -23,7 +23,8 @@
     "docker:up": "docker-compose up",
     "docker:start": "docker-compose start",
     "docker:down": "docker-compose down",
-    "db:create": "node sql/creation.js"
+    "db:create": "node scripts/creation.js",
+    "db:seeds": "node scripts/seeds.js"
   },
   "repository": {
     "type": "git",
